Mod: rm world offset, geyser dormancy info in int

+++ mod/_WorldGenStateCapture/WorldStateData/WorldPOIs/MapGeyser.cs
@@ -8,7 +8,7 @@
 		public int idleTime;
 		public int eruptionTime;
 
-		public float dormancyCycles;
-		public float activeCycles;
+		public int dormancyCycles;
+		public int activeCycles;
 	}
 }
